MT#60551 Fix docker run instructions
The run instructions were still pointing to the bullseye image
instead of the bookworm one.

Additionally the command to call once connected to the docker
bash has been updated: it is not necessary to move to "code" folder
since we are already there.

Finally executable permissions has been given to t/run_web_e2e
file.

Change-Id: Iaeb19713309095136a5149f56ceae6b7bf8a6103